About the Project
Ludo game implemented in C with the use of sockets and threads. 
For the purposes of semestral project it runs on a server and all the players can join and play there as clients up to 4 people are allowed.
When a player who manages to be a server as well leaves, the game stops and all the players are automatically disconnected.